requirements.txt
chardet==5.2.0
python-dotenv==1.0.1
